+++ circuits/core/src/algorithms/poseidon/ecies.rs
@@ -105,7 +105,7 @@ mod tests {
     };
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 10;
+    const ITERATIONS: u64 = 10;
     const MESSAGE: &str = "ECIESCircuit0";
 
     fn check_encode_decode(mode: Mode) {
