+++ circuits/types/field/src/equal.rs
@@ -161,7 +161,7 @@ mod tests {
     use snarkvm_circuits_environment::Circuit;
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 200;
+    const ITERATIONS: u64 = 200;
 
     fn check_is_equal(name: &str, expected: bool, a: &Field<Circuit>, b: &Field<Circuit>) {
         Circuit::scope(name, || {
