+++ circuits/types/field/src/sub.rs
@@ -86,7 +86,7 @@ mod tests {
     use snarkvm_circuits_environment::Circuit;
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 10_000;
+    const ITERATIONS: u64 = 10_000;
 
     fn check_sub(name: &str, expected: &<Circuit as Environment>::BaseField, a: &Field<Circuit>, b: &Field<Circuit>) {
         Circuit::scope(name, || {
