+++ circuits/types/group/src/mul.rs
@@ -187,17 +187,17 @@ mod tests {
     use snarkvm_curves::ProjectiveCurve;
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 10;
+    const ITERATIONS: u64 = 10;
 
     fn check_mul(
         name: &str,
         expected: &<Circuit as Environment>::Affine,
         a: &Group<Circuit>,
         b: &Scalar<Circuit>,
-        num_constants: usize,
-        num_public: usize,
-        num_private: usize,
-        num_constraints: usize,
+        num_constants: u64,
+        num_public: u64,
+        num_private: u64,
+        num_constraints: u64,
     ) {
         Circuit::scope(name, || {
             let candidate = a * b;
@@ -212,10 +212,10 @@ mod tests {
         expected: &<Circuit as Environment>::Affine,
         a: &Group<Circuit>,
         b: &Scalar<Circuit>,
-        num_constants: usize,
-        num_public: usize,
-        num_private: usize,
-        num_constraints: usize,
+        num_constants: u64,
+        num_public: u64,
+        num_private: u64,
+        num_constraints: u64,
     ) {
         Circuit::scope(name, || {
             let mut candidate = a.clone();
@@ -287,8 +287,8 @@ mod tests {
             let scalar: <Circuit as Environment>::ScalarField = UniformRand::rand(&mut test_rng());
 
             let (num_constant, num_private, num_constraints) = {
-                const MODULUS_BITS: usize = 251;
-                let num_nonzero_bits = scalar.to_repr().to_biguint().bits() as usize;
+                const MODULUS_BITS: u64 = 251;
+                let num_nonzero_bits = scalar.to_repr().to_biguint().bits();
                 let num_leading_zero_bits = MODULUS_BITS - num_nonzero_bits;
 
                 let num_constant = 2
@@ -327,8 +327,8 @@ mod tests {
             let scalar: <Circuit as Environment>::ScalarField = UniformRand::rand(&mut test_rng());
 
             let (num_constant, num_private, num_constraints) = {
-                const MODULUS_BITS: usize = 251;
-                let num_nonzero_bits = scalar.to_repr().to_biguint().bits() as usize;
+                const MODULUS_BITS: u64 = 251;
+                let num_nonzero_bits = scalar.to_repr().to_biguint().bits();
                 let num_leading_zero_bits = MODULUS_BITS - num_nonzero_bits;
 
                 let num_constant = 2
